build: add debian gcc8 build
